refactor(allocator): extract session-metrics routes into routes/metrics.py
Moves post_session_metrics, _session_metrics_view_model,
admin_session_metrics, get_session_metrics_summary_json, and
export_metrics out of main.py into a new metrics blueprint, hoisting
export_metrics's function-local csv/io/datetime imports to module
scope. This was the last route move: main.py now has zero @app.route
decorators and is purely the wiring-and-state module.
